Megakkart the largest ecommerce platform in the country. Owing to the festive season Megakkart is planning a ‘Super Festive Sale’ where it wants to achieve a  Gross merchandise value,GMV sales of INR 5000 crores.
Megakkart the largest ecommerce platform in the country. Owing to the festive season Megakkart is planning a ‘Super Festive Sale’ where it wants to achieve a  Gross merchandise value,GMV sales of INR 5000 crores.
The company plans to work on various hooks for the customers who are at a risk of not making a purchase. The analysis and the machine learning model on the last year data will help Megakkart product team to plan these hooks, along with identifying the profitable customers.